Modify the data type used for address offsets



# message auto-generated for no-merge-commit merge:
!332 merge fa_251113 into master

Modify the data type used for address offsets

Created-by: pencho
Commit-by: pencho
Merged-by: turing_project1
Description: ## 描述
Modify the data type used for address offsets

## 关联的Issue
关联Issue #154

## 测试
算子编译

## 文档更新
无

## 类型标签
<!--  [x] 表示选中 -->
- [ ] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [X] 其他，请描述：修改地址偏移使用的数据类型为int64_t




See merge request: cann/ops-transformer!332

+++ attention/flash_attention_score/op_kernel/flash_attention_score_bn2gs1s2_b.h
@@ -1096,7 +1096,7 @@ FlashAttentionScoreBn2gs1s2B<FA_BN2GS1S2B_FUNCTION_PARAMS_TEMPLATE>::ProcessVec2
         int64_t biN2GSumOffset = extraInfo.biN2GoIdx * this->s1Size * fp32BaseSize;
         int64_t sumGmOffset = bOffset + biN2GSumOffset;
         for (int64_t s1oIdx = 0; s1oIdx < extraInfo.s1Vec2OuterSize; ++s1oIdx) {
-            int sumGmRealOffset = 0;
+            int64_t sumGmRealOffset = 0;
             extraInfo.s1Vec2BaseSize = this->s1Vec2BaseSize;
             if (s1oIdx == extraInfo.s1Vec2OuterSize - 1) {
                 extraInfo.s1Vec2BaseSize = extraInfo.s1Vec2BaseTailSize;
